Update README.md
Greetings from the Netlify Docs team!

I’m proposing a small update to your README. Build Plugins are now out of beta, and we want to encourage plugins that are included in the Plugins directory to include UI-based installation instructions. 

Feel free to edit and merge as you see fit.